refactor: Centralize App and Permission Configuration
This commit introduces several refactoring improvements to centralize application configurations and enhance the permission handling system.

Furthermore, it fix the imports for some files
