+++ app/src/androidTest/java/com/github/warnastrophy/core/service/HazardTrackingServiceTest.kt
@@ -1,8 +1,8 @@
 package com.github.warnastrophy.core.service
 
-import com.github.warnastrophy.core.data.repository.usecase.RefreshHazardsIfMovedUseCase
 import com.github.warnastrophy.core.data.service.HazardTrackingService
-import com.github.warnastrophy.core.model.Location
+import com.github.warnastrophy.core.domain.model.Location
+import com.github.warnastrophy.core.domain.usecase.RefreshHazardsIfMovedUseCase
 import com.github.warnastrophy.core.ui.map.GpsServiceMock
 import com.google.android.gms.maps.model.LatLng
 import io.mockk.clearAllMocks
